+++ hipcub/include/hipcub/hipcub.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2020 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2018-2024 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -28,9 +28,7 @@
 /// Meta-header to include all hipCUB APIs.
 
 // Meta configuration for hipCUB
-#include "config.hpp"
-// Version
-#include "hipcub_version.hpp"
+#include "config.hpp" // IWYU pragma: export
 
 #ifdef __HIP_PLATFORM_AMD__
     #include "backend/rocprim/hipcub.hpp"